+++ demo/utils/chains.ts
@@ -158,5 +158,14 @@ export default [
   },
   {
     ...reservoirChains.frameTestnet
-  }
+  },
+  {
+    ...reservoirChains.blastSepolia
+  },
+  {
+    ...reservoirChains.apexPop
+  },
+  {
+    ...reservoirChains.blast
+  },
 ]